refactor[resourcesbot]: add `modules` directory
