+++ pywikitools/resourcesbot/modules/export_pdf.py
@@ -7,7 +7,7 @@ from typing import Final, Optional
 from pywikitools.fortraininglib import ForTrainingLib
 from pywikitools.resourcesbot.changes import ChangeLog
 from pywikitools.resourcesbot.data_structures import FileInfo, LanguageInfo
-from pywikitools.resourcesbot.post_processing import LanguagePostProcessor
+from pywikitools.resourcesbot.modules.post_processing import LanguagePostProcessor
 
 
 class ExportPDF(LanguagePostProcessor):
